requirements.txt
numpy
torch
faiss-gpu
sentence-transformers
scikit-learn
tqdm
transformers
pandas
datasets
rank-bm25
huggingface-hub
cohere